requirements.txt
asgiref==3.10.0
autoflake==2.3.1
black==25.9.0
click==8.3.0
colorama==0.4.6
cssbeautifier==1.15.4
Django==5.2.7
django-environ==0.12.0
djangorestframework==3.16.1
djlint==1.36.4
drf-yasg==1.21.11
EditorConfig==0.17.1
flake8==7.3.0
inflection==0.5.1
isort==7.0.0
jsbeautifier==1.15.4
json5==0.12.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.2.0
